Assignment-2/Task-1/requirements.txt
Flask==3.0.0
mysql-connector-python==8.3.0
